General project cleanup & converted CLI command system into an attribute/reflection based one to reduce code duplication and multiple sources of truth.
Added the ability to use a .accesstoken file instead of providing the access token on every invocation. The file is a fallback.

Added a CopyTags command, allowing you to clone the tags of any arbitrary local Git repository onto the specified GitLab project. Will only work if the two repos have the same git histories (it tags by commits)