feature: IHttpClientProxy
Identical functionally to an HttpClient, it is however comprised of a user-definable `SendAsync`. The main use case of this is to simply forward it to HttpClient and log information about the request, such as HTTP method, execution time, etc.
This interface provides no other functionality onto HttpClient other than simple endpoint logging or request middlewares. It should be parameter type & order compatible with the SendAsync, GetAsync, PutAsync, etc methods directly on HttpClient.